Qt 6.2 LTS.

+++ Source/dooble.cc
@@ -2205,9 +2205,9 @@ void dooble::print_preview(QPrinter *printer, dooble_charts *chart)
   auto view = chart->view();
 #if (QT_VERSION >= QT_VERSION_CHECK(5, 15, 0))
   auto xscale = printer->pageLayout().fullRectPixels(printer->resolution()).
-    width() / view->width();
+    width() / static_cast<double> (view->width());
   auto yscale = printer->pageLayout().fullRectPixels(printer->resolution()).
-    height() / view->height();
+    height() / static_cast<double> (view->height());
 #else
   auto xscale = printer->pageRect().width() /
     static_cast<double> (view->width());
